autenticacion con auth.js (ex next-auth)

+++ 03-admin-todos/src/app/api/todos/route.ts
@@ -1,4 +1,6 @@
 import * as yup from 'yup'
+
+import { getUserSession } from '@/auth/actions/auth-actions'
 import prisma from '@/lib/prisma'
 
 export async function GET(request: Request) {
@@ -16,8 +18,10 @@ const postSchema = yup.object({
 
 export async function POST(request: Request) {
   try {
+    const user = await getUserSession()
+    if (!user) throw new Error('No autorizado')
     const { complete, description } = await postSchema.validate(await request.json())
-    const todo = await prisma.todo.create({ data: { complete, description } })
+    const todo = await prisma.todo.create({ data: { complete, description, userId: user.id } })
     return Response.json({ todo }, { status: 201 })
   } catch (error) {
     return Response.json({ error }, { status: 400 })
@@ -26,7 +30,9 @@ export async function POST(request: Request) {
 
 export async function DELETE() {
   try {
-    await prisma.todo.deleteMany({ where: { complete: true } })
+    const user = await getUserSession()
+    if (!user) throw new Error('No autorizado')
+    await prisma.todo.deleteMany({ where: { complete: true, userId: user.id } })
     return Response.json({})
   } catch (error) {
     return Response.json({ error }, { status: 400 })
